refactor(ui): split SessionSidebar.tsx god object into modules
